pyproject.toml
[project]
name = "agentmesh"
version = "0.1.0"
description = "MCP-Native Multi-Agent Orchestrator with Live Mission Control"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.0.0",
    "websockets>=12.0",
    "langgraph>=0.2.0",
    "fastmcp>=1.0.0",
    "google-genai>=1.0.0",
    "groq>=0.5.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "anyio>=4.0.0",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "python-jose[cryptography]>=3.3",
    "cryptography>=42.0",
    "psycopg2-binary>=2.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.14.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes (unused imports, undefined names)
    "B",   # flake8-bugbear (common bugs)
    "I",   # isort (import order)
    "UP",  # pyupgrade (modern python syntax)
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # FastAPI Depends() in function args is the correct pattern
    "E402",  # module-level imports not at top (intentional in alembic/env.py)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
